service/requirements/common.txt
Django==1.11
birdy==0.3.2
pymongo==3.6.1
celery[redis]==4.1.0
django-environ==0.4.4
djangorestframework==3.7.7
